refact(geo.service): move to core module
